# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 兜行 Douxing API
  description: |
    兜行平台 REST API（`packages/server`）。

    - 统一响应：`{ code, message, messageKey?, data }`，`code === 0` 为成功
    - 鉴权：Header `Authorization: Bearer <Access Token JWT>`；过期后客户端用 Refresh Token 调 `POST /api/auth/refresh`（见 [双Token认证与无感刷新.md](./双Token认证与无感刷新.md)）
    - 国际化：Header `Accept-Language: zh-CN | en-US`
    - 人类可读说明见 [API接口文档.md](./API接口文档.md)
  version: 1.4.1
  contact:
    name: 兜行项目
servers:
  - url: http://localhost:3000
    description: 本地开发
  - url: https://api.yxtao.site
    description: 生产环境
...
security:
  - bearerAuth: []
paths:
...
  /api/analytics/overview:
    get:
      tags: [数据分析]
      summary: 核心指标概览（管理员）
      operationId: analyticsOverview
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AnalyticsOverview'
        '403':
          description: 非管理员
  /api/analytics/trends:
    get:
      tags: [数据分析]
      summary: 按日趋势（管理员）
      operationId: analyticsTrends
      parameters:
        - name: days
          in: query
          description: 天数，1～90，默认 30
          schema:
            type: integer
            minimum: 1
            maximum: 90
            default: 30
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/AnalyticsDailyPoint'
        '403':
          description: 非管理员
  /api/analytics/top-cities:
    get:
      tags: [数据分析]
      summary: 打卡城市排行（管理员）
      operationId: analyticsTopCities
      parameters:
        - name: limit
          in: query
          description: Top N，1～50，默认 10
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/AnalyticsCityRankItem'
        '403':
          description: 非管理员
  /api/analytics/funnel:
    get:
      tags: [数据分析]
      summary: 旅程漏斗（管理员，DT3）
      operationId: analyticsFunnel
      parameters:
        - name: days
          in: query
          description: 天数，1～90，默认 30
          schema:
            type: integer
            minimum: 1
            maximum: 90
            default: 30
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/AnalyticsFunnelStep'
        '403':
          description: 非管理员
  /api/analytics/geo/distribution:
    get:
      tags: [数据分析]
      summary: 地理分布与热力点（管理员，DT5）
      operationId: analyticsGeoDistribution
      parameters:
        - name: days
          in: query
          description: 天数，1～90，默认 30
          schema:
            type: integer
            minimum: 1
            maximum: 90
            default: 30
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AnalyticsGeoDistribution'
        '403':
          description: 非管理员
  /api/analytics/geo/flows:
    get:
      tags: [数据分析]
      summary: 城际 OD 流动（管理员，DT5）
      operationId: analyticsGeoFlows
      parameters:
        - name: days
          in: query
          description: 天数，1～90，默认 30
          schema:
            type: integer
            minimum: 1
            maximum: 90
            default: 30
        - name: limit
          in: query
          description: Top N，1～50，默认 20
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/AnalyticsGeoFlow'
        '403':
          description: 非管理员
  /api/analytics/system-resources:
    get:
      tags: [数据分析]
      summary: 系统资源统计（远程 Git，管理员）
      description: |
        浅克隆远程仓库后分类统计代码/组件/页面/模块/文档与提交趋势。
        不扫描部署机本机业务工作区。详见 docs/系统资源统计.md。
      operationId: analyticsSystemResources
      parameters:
        - name: refresh
          in: query
          description: 为 true/1 时强制重新 fetch 远程并刷新缓存
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SystemResourcesStats'
        '403':
          description: 非管理员或无 data:analytics:view
        '500':
          description: 远程 Git 拉取或聚合失败（messageKey api.systemResourcesGitFailed / api.systemResourcesFailed）
  /api/analytics/events:
    post:
      tags: [数据分析]
      summary: 写入埋点事件（管理员）
      operationId: analyticsTrackEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrackAnalyticsEventRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          ok:
                            type: boolean
                            example: true
        '403':
          description: 非管理员
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "登录后获得的 Access Token（JWT），填入 Authorization: Bearer <token>"
...
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: 成功
        messageKey:
          type: string
          example: api.ok
        data:
          nullable: true
...
    AnalyticsOverview:
      type: object
      properties:
        users:
          type: object
          properties:
            total: { type: integer }
            newToday: { type: integer }
            newLast7Days: { type: integer }
        routes:
          type: object
          properties:
            total: { type: integer }
            publicTotal: { type: integer }
            newLast7Days: { type: integer }
        orders:
          type: object
          properties:
            total: { type: integer }
            paidTotal: { type: integer }
            newLast7Days: { type: integer }
        checkins:
          type: object
          properties:
            total: { type: integer }
            approvedTotal: { type: integer }
            newLast7Days: { type: integer }
        planSessions:
          type: object
          properties:
            total: { type: integer }
            newLast7Days: { type: integer }
        generatedAt: { type: string, format: date-time }
    SystemResourcesStats:
      type: object
      description: GET /api/analytics/system-resources 响应体（远程 Git 快照）
      properties:
        source:
          type: string
          enum: [git-remote]
        remoteUrl: { type: string, description: 脱敏后的远程地址 }
        branch: { type: string }
        commitSha: { type: string }
        commitAt: { type: string, format: date-time }
        apiRouteCount: { type: integer }
        generatedAt: { type: string, format: date-time }
        projectRoot:
          type: string
          deprecated: true
          description: 兼容字段，优先使用 remoteUrl
        packages:
          type: array
          items: { $ref: '#/components/schemas/SystemResourcePackageStat' }
        fileTypes:
          type: array
          items: { $ref: '#/components/schemas/SystemResourceFileTypeStat' }
        docs:
          type: array
          items: { $ref: '#/components/schemas/SystemResourceDocStat' }
        componentsByApp:
          type: array
          items: { $ref: '#/components/schemas/SystemResourcePackageStat' }
        pagesByApp:
          type: array
          items: { $ref: '#/components/schemas/SystemResourcePackageStat' }
        modules:
          type: array
          items: { $ref: '#/components/schemas/SystemResourcePackageStat' }
        plugins:
          type: array
          items: { $ref: '#/components/schemas/SystemResourcePackageStat' }
        dependencies:
          type: array
          items: { $ref: '#/components/schemas/SystemResourcePackageStat' }
        commitTrends:
          type: array
          items: { $ref: '#/components/schemas/SystemResourceCommitDayStat' }
        moduleChurn:
          type: array
          items: { $ref: '#/components/schemas/SystemResourceModuleChurnStat' }
        overview:
          type: array
          items: { $ref: '#/components/schemas/SystemResourceMetricItem' }
        i18nKeys:
          type: array
          items: { $ref: '#/components/schemas/SystemResourceMetricItem' }
    SystemResourcePackageStat:
      type: object
      properties:
        name: { type: string }
        path: { type: string }
        fileCount: { type: integer }
        lineCount: { type: integer }
    SystemResourceFileTypeStat:
      type: object
      properties:
        ext: { type: string }
        fileCount: { type: integer }
        lineCount: { type: integer }
    SystemResourceDocStat:
      type: object
      properties:
        path: { type: string }
        fileCount: { type: integer }
        lineCount: { type: integer }
    SystemResourceCommitDayStat:
      type: object
      properties:
        date: { type: string, example: '2026-07-17' }
        insertions: { type: integer }
        deletions: { type: integer }
        sourceInsertions: { type: integer }
        sourceDeletions: { type: integer }
        commits: { type: integer }
        totalLines: { type: integer }
    SystemResourceModuleChurnStat:
      type: object
      properties:
        name: { type: string }
        insertions: { type: integer }
        deletions: { type: integer }
    SystemResourceMetricItem:
      type: object
      properties:
        labelKey: { type: string }
        value: { type: number }
        descKey: { type: string }
    AnalyticsDailyPoint:
      type: object
      properties:
        date: { type: string, example: '2026-06-09' }
        users: { type: integer }
        routes: { type: integer }
        orders: { type: integer }
        checkins: { type: integer }
        planSessions: { type: integer }
    AnalyticsCityRankItem:
      type: object
      properties:
        cityCode: { type: string }
        checkinCount: { type: integer }
    AnalyticsFunnelStep:
      type: object
      properties:
        stepKey: { type: string }
        eventName: { type: string }
        count: { type: integer }
        rateFromFirst: { type: number, nullable: true }
        rateFromPrev: { type: number, nullable: true }
    AnalyticsProvinceGeoStat:
      type: object
      properties:
        provinceCode: { type: string }
        geoName: { type: string }
        adcode: { type: string }
        checkinCount: { type: integer }
    AnalyticsCityGeoStat:
      type: object
      properties:
        cityCode: { type: string }
        cityName: { type: string }
        latitude: { type: number }
        longitude: { type: number }
        checkinCount: { type: integer }
    AnalyticsHeatPoint:
      type: object
      properties:
        latitude: { type: number }
        longitude: { type: number }
        weight: { type: number }
    AnalyticsGeoDistribution:
      type: object
      properties:
        provinces:
          type: array
          items:
            $ref: '#/components/schemas/AnalyticsProvinceGeoStat'
        cities:
          type: array
          items:
            $ref: '#/components/schemas/AnalyticsCityGeoStat'
        heatPoints:
          type: array
          items:
            $ref: '#/components/schemas/AnalyticsHeatPoint'
        scopeDays: { type: integer }
        effectiveScope:
          type: string
          enum: [window, allTime]
        generatedAt: { type: string, format: date-time }
    AnalyticsGeoFlow:
      type: object
      properties:
        fromCityCode: { type: string }
        toCityCode: { type: string }
        fromName: { type: string }
        toName: { type: string }
        fromLatitude: { type: number }
        fromLongitude: { type: number }
        toLatitude: { type: number }
        toLongitude: { type: number }
        count: { type: integer }
...
    TrackAnalyticsEventRequest:
      type: object
      required: [eventName]
      properties:
        eventName:
          type: string
          minLength: 1
          maxLength: 64
        eventCategory:
          type: string
          enum: [business, behavior, system]
        userId:
          type: integer
          minimum: 1
        sessionId:
          type: string
          maxLength: 64
        properties:
          type: object
          additionalProperties: true
        source:
          type: string
          enum: [server, mobile, pc, web]
        occurredAt:
          type: string
          format: date-time
...